templane-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "templane-python"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["pyyaml>=6.0", "jinja2>=3.1"]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.hatch.build.targets.wheel]
packages = [
  "src/templane_core",
  "src/templane_adapter_html",
  "src/templane_adapter_yaml",
  "src/jinja_templane",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
